requirements.txt
Flask>=2.3.0
pymongo>=4.5.0
Werkzeug>=2.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
catboost>=1.2.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
dnspython>=2.4.0
setuptools>=65.0.0
Flask-Mail>=0.9.1